+++ properties/AddingOutsideOfTransactionIsNotAllowed.php
@@ -41,7 +41,8 @@ final class AddingOutsideOfTransactionIsNotAllowed implements Property
         $assert->throws(
             static fn() => $manager
                 ->repository(User::class)
-                ->put($user),
+                ->put($user)
+                ->unwrap(),
             \LogicException::class,
             'Mutation outside of a transaction',
         );