+++ properties/EffectChildAddOnAggregate.php
@@ -18,7 +18,6 @@ use Innmind\Specification\{
     Comparator,
     Sign,
 };
-use Innmind\Immutable\Either;
 use Fixtures\Innmind\TimeContinuum\PointInTime;
 
 /**
@@ -57,11 +56,10 @@ final class EffectChildAddOnAggregate implements Property
     {
         $user = User::new($this->createdAt, $this->name);
         $manager->transactional(
-            static fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->put($user),
-            ),
+            static fn() => $manager
+                ->repository(User::class)
+                ->put($user)
+                ->either(),
         );
         $id = $user->id()->toString();
         unset($user); // to make sure there is no in memory cache somewhere
@@ -73,16 +71,15 @@ final class EffectChildAddOnAggregate implements Property
         );
 
         $manager->transactional(
-            fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->effect(
-                        Effect::collection('addresses')->add(
-                            User\Address::new($this->address),
-                        ),
-                        $specification,
+            fn() => $manager
+                ->repository(User::class)
+                ->effect(
+                    Effect::collection('addresses')->add(
+                        User\Address::new($this->address),
                     ),
-            ),
+                    $specification,
+                )
+                ->either(),
         );
 
         $manager