+++ properties/EffectChildRemoveOnAllAggregates.php
@@ -17,7 +17,6 @@ use Innmind\BlackBox\{
     Property,
     Runner\Assert,
 };
-use Innmind\Immutable\Either;
 use Fixtures\Innmind\TimeContinuum\PointInTime;
 
 /**
@@ -68,25 +67,23 @@ final class EffectChildRemoveOnAllAggregates implements Property
     {
         $user = User::new($this->createdAt, $this->name);
         $manager->transactional(
-            static fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->put($user),
-            ),
+            static fn() => $manager
+                ->repository(User::class)
+                ->put($user)
+                ->either(),
         );
         unset($user); // to make sure there is no in memory cache somewhere
 
         $address = $this->prefix.$this->suffix;
         $manager->transactional(
-            static fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->effect(
-                        Effect::collection('addresses')->add(
-                            User\Address::new($address),
-                        ),
+            static fn() => $manager
+                ->repository(User::class)
+                ->effect(
+                    Effect::collection('addresses')->add(
+                        User\Address::new($address),
                     ),
-            ),
+                )
+                ->either(),
         );
 
         $manager
@@ -112,19 +109,18 @@ final class EffectChildRemoveOnAllAggregates implements Property
         };
 
         $manager->transactional(
-            fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->effect(
-                        Effect::collection('addresses')->remove(
-                            Comparator\Property::of(
-                                'value',
-                                $this->sign,
-                                $value,
-                            ),
+            fn() => $manager
+                ->repository(User::class)
+                ->effect(
+                    Effect::collection('addresses')->remove(
+                        Comparator\Property::of(
+                            'value',
+                            $this->sign,
+                            $value,
                         ),
                     ),
-            ),
+                )
+                ->either(),
         );
 
         $manager