+++ properties/EffectEntityPropertiesOnAllAggregates.php
@@ -13,7 +13,6 @@ use Innmind\BlackBox\{
     Property,
     Runner\Assert,
 };
-use Innmind\Immutable\Either;
 use Fixtures\Innmind\TimeContinuum\PointInTime;
 
 /**
@@ -54,30 +53,28 @@ final class EffectEntityPropertiesOnAllAggregates implements Property
     {
         $user = User::new($this->createdAt, $this->name);
         $manager->transactional(
-            static fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->put($user),
-            ),
+            static fn() => $manager
+                ->repository(User::class)
+                ->put($user)
+                ->either(),
         );
         unset($user); // to make sure there is no in memory cache somewhere
 
         $manager->transactional(
-            fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->effect(
-                        Effect::entity('mainAddress')->properties(
-                            Effect::property('value')
-                                ->assign($this->address)
-                                ->and(
-                                    Effect::property('enabled')->assign(
-                                        $this->enabled,
-                                    ),
+            fn() => $manager
+                ->repository(User::class)
+                ->effect(
+                    Effect::entity('mainAddress')->properties(
+                        Effect::property('value')
+                            ->assign($this->address)
+                            ->and(
+                                Effect::property('enabled')->assign(
+                                    $this->enabled,
                                 ),
-                        ),
+                            ),
                     ),
-            ),
+                )
+                ->either(),
         );
 
         $manager