+++ properties/EffectOptionalNothingOnAllAggregates.php
@@ -13,7 +13,6 @@ use Innmind\BlackBox\{
     Property,
     Runner\Assert,
 };
-use Innmind\Immutable\Either;
 use Fixtures\Innmind\TimeContinuum\PointInTime;
 
 /**
@@ -54,23 +53,19 @@ final class EffectOptionalNothingOnAllAggregates implements Property
             $this->address,
         );
         $manager->transactional(
-            static fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->put($user),
-            ),
+            static fn() => $manager
+                ->repository(User::class)
+                ->put($user)
+                ->either(),
         );
         $id = $user->id()->toString();
         unset($user); // to make sure there is no in memory cache somewhere
 
         $manager->transactional(
-            static fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->effect(
-                        Effect::optional('billingAddress')->nothing(),
-                    ),
-            ),
+            static fn() => $manager
+                ->repository(User::class)
+                ->effect(Effect::optional('billingAddress')->nothing())
+                ->either(),
         );
 
         $manager