+++ properties/EffectOptionalPropertiesOnAllAggregates.php
@@ -13,7 +13,6 @@ use Innmind\BlackBox\{
     Property,
     Runner\Assert,
 };
-use Innmind\Immutable\Either;
 use Fixtures\Innmind\TimeContinuum\PointInTime;
 
 /**
@@ -58,25 +57,23 @@ final class EffectOptionalPropertiesOnAllAggregates implements Property
             $this->address,
         );
         $manager->transactional(
-            static fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->put($user),
-            ),
+            static fn() => $manager
+                ->repository(User::class)
+                ->put($user)
+                ->either(),
         );
         $id = $user->id()->toString();
         unset($user); // to make sure there is no in memory cache somewhere
 
         $manager->transactional(
-            fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->effect(
-                        Effect::optional('billingAddress')->properties(
-                            Effect::property('value')->assign($this->newAddress),
-                        ),
+            fn() => $manager
+                ->repository(User::class)
+                ->effect(
+                    Effect::optional('billingAddress')->properties(
+                        Effect::property('value')->assign($this->newAddress),
                     ),
-            ),
+                )
+                ->either(),
         );
 
         $manager