+++ properties/EffectPropertyOnAllAggregates.php
@@ -13,7 +13,6 @@ use Innmind\BlackBox\{
     Property,
     Runner\Assert,
 };
-use Innmind\Immutable\Either;
 use Fixtures\Innmind\TimeContinuum\PointInTime;
 
 /**
@@ -57,22 +56,20 @@ final class EffectPropertyOnAllAggregates implements Property
     {
         $user = User::new($this->createdAt, $this->name);
         $manager->transactional(
-            static fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->put($user),
-            ),
+            static fn() => $manager
+                ->repository(User::class)
+                ->put($user)
+                ->either(),
         );
         unset($user); // to make sure there is no in memory cache somewhere
 
         $manager->transactional(
-            fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->effect(Effect::property('name')->assign(
-                        $this->newName,
-                    )),
-            ),
+            fn() => $manager
+                ->repository(User::class)
+                ->effect(Effect::property('name')->assign(
+                    $this->newName,
+                ))
+                ->either(),
         );
 
         $manager