+++ properties/MatchingSortEntity.php
@@ -72,8 +72,8 @@ final class MatchingSortEntity implements Property
         $repository = $manager->repository(User::class);
         $manager->transactional(
             static function() use ($repository, $user1, $user2) {
-                $repository->put($user1);
-                $repository->put($user2);
+                $repository->put($user1)->unwrap();
+                $repository->put($user2)->unwrap();
 
                 return Either::right(null);
             },