+++ properties/RemoveWhereEntity.php
@@ -66,8 +66,8 @@ final class RemoveWhereEntity implements Property
         $repository = $manager->repository(User::class);
         $manager->transactional(
             static function() use ($repository, $user1, $user2) {
-                $repository->put($user1);
-                $repository->put($user2);
+                $repository->put($user1)->unwrap();
+                $repository->put($user2)->unwrap();
 
                 return Either::right(null);
             },
@@ -82,7 +82,7 @@ final class RemoveWhereEntity implements Property
                 $repository->remove(MainAddress::of(
                     Sign::equality,
                     $this->name1,
-                ));
+                ))->unwrap();
 
                 return Either::right(null);
             },