+++ src/Adapter/Elasticsearch/Repository.php
@@ -34,6 +34,8 @@ use Innmind\Immutable\{
     Sequence,
     Maybe,
     Map,
+    Attempt,
+    SideEffect,
 };
 
 /**
@@ -155,9 +157,9 @@ final class Repository implements RepositoryInterface, Effectful
     }
 
     #[\Override]
-    public function add(Aggregate $data): void
+    public function add(Aggregate $data): Attempt
     {
-        $_ = ($this->http)(Request::of(
+        return ($this->http)(Request::of(
             self::url(
                 $this->url,
                 $this->path,
@@ -170,16 +172,17 @@ final class Repository implements RepositoryInterface, Effectful
                 ContentType::of(new MediaType('application', 'json')),
             ),
             ($this->encode)($data),
-        ))->match(
-            static fn() => null,
-            static fn() => throw new \RuntimeException('Unable to persist the aggregate'),
-        );
+        ))
+            ->attempt(
+                static fn() => new \RuntimeException('Unable to persist the aggregate'),
+            )
+            ->map(static fn() => SideEffect::identity());
     }
 
     #[\Override]
-    public function update(Diff $data): void
+    public function update(Diff $data): Attempt
     {
-        $_ = ($this->http)(Request::of(
+        return ($this->http)(Request::of(
             self::url(
                 $this->url,
                 $this->path,
@@ -192,17 +195,18 @@ final class Repository implements RepositoryInterface, Effectful
                 ContentType::of(new MediaType('application', 'json')),
             ),
             ($this->encode)($data),
-        ))->match(
-            static fn() => null,
-            static fn() => throw new \RuntimeException('Unable to update the aggregate'),
-        );
+        ))
+            ->attempt(
+                static fn() => new \RuntimeException('Unable to update the aggregate'),
+            )
+            ->map(static fn() => SideEffect::identity());
     }
 
     #[\Override]
     public function effect(
         Effect\Normalized $effect,
         ?Specification $specification,
-    ): void {
+    ): Attempt {
         $payload = [
             'script' => ($this->script)($effect),
         ];
@@ -211,7 +215,7 @@ final class Repository implements RepositoryInterface, Effectful
             $payload['query'] = ($this->query)($specification);
         }
 
-        $_ = ($this->http)(Request::of(
+        return ($this->http)(Request::of(
             self::url(
                 $this->url,
                 $this->path,
@@ -223,16 +227,17 @@ final class Repository implements RepositoryInterface, Effectful
                 ContentType::of(new MediaType('application', 'json')),
             ),
             Content::ofString(Json::encode($payload)),
-        ))->match(
-            static fn() => null,
-            static fn() => throw new \RuntimeException('Unable to update multiple aggregates'),
-        );
+        ))
+            ->attempt(
+                static fn() => new \RuntimeException('Unable to update multiple aggregates'),
+            )
+            ->map(static fn() => SideEffect::identity());
     }
 
     #[\Override]
-    public function remove(Aggregate\Id $id): void
+    public function remove(Aggregate\Id $id): Attempt
     {
-        $_ = ($this->http)(Request::of(
+        return ($this->http)(Request::of(
             self::url(
                 $this->url,
                 $this->path,
@@ -241,16 +246,18 @@ final class Repository implements RepositoryInterface, Effectful
             ),
             Method::delete,
             ProtocolVersion::v11,
-        ))->match(
-            static fn() => null,
-            static fn() => null,
-        );
+        ))
+            ->attempt(
+                static fn() => new \RuntimeException('Failed to remove aggregate'),
+            )
+            ->recover(static fn() => Attempt::result(SideEffect::identity()))
+            ->map(static fn() => SideEffect::identity());
     }
 
     #[\Override]
-    public function removeAll(Specification $specification): void
+    public function removeAll(Specification $specification): Attempt
     {
-        $_ = ($this->http)(Request::of(
+        return ($this->http)(Request::of(
             self::url(
                 $this->url,
                 $this->path,
@@ -264,10 +271,11 @@ final class Repository implements RepositoryInterface, Effectful
             Content::ofString(Json::encode([
                 'query' => ($this->query)($specification),
             ])),
-        ))->match(
-            static fn() => null,
-            static fn() => throw new \RuntimeException('Unable to remove multiple aggregates'),
-        );
+        ))
+            ->attempt(
+                static fn() => new \RuntimeException('Unable to remove multiple aggregates'),
+            )
+            ->map(static fn() => SideEffect::identity());
     }
 
     #[\Override]